refactor(UnionBan): refactoring UnionBan module
