+++ src/main/java/fun/blue_millennium/data/UnionBan/LocalProcess/UnionBanDataGet.java
@@ -1,4 +1,4 @@
-package fun.blue_millennium.data.UnionBan;
+package fun.blue_millennium.data.UnionBan.LocalProcess;
 
 import com.google.gson.Gson;
 import com.google.gson.reflect.TypeToken;
